requirements.txt
django==2.2.9
django-extensions==2.2.6
django-debug-toolbar
pillow
